調整排版以及修改格式錯誤顯示

+++ Delete.php
@@ -1,10 +1,16 @@
 <?php require_once('connect.php'); ?>
 <?php 
 //刪除通訊錄資料
-    $sqlDelete = "Delete from ListView where Id='" . $_POST["Id"] . "'";
-    $conn->query($sqlDelete);
-    echo "資料已刪除";
-
+    if(preg_match("/^[0-9]+$/", $_POST["Id"])){
+        $sqlDelete = "Delete from ListView where Id='" . $_POST["Id"] . "'";
+        $conn->query($sqlDelete);
+        //echo "資料已刪除";
+        header('Location: index.php'); 
+    }else{
+        echo "<script>alert('資料輸入有誤 刪除失敗')</script>";
+        header('Location: index.php'); 
+    }
+    
 ?>
 
 <html>
@@ -13,7 +19,6 @@
 </head>
 
 <body>
-        <input type ="button" onclick="javascript:location.href='index.php'" value="Home"></input>   
 </body>
 
 </html>
